refactor: 폴더구조 개변(lib활용)
