+++ src/lib/hook/index.ts
@@ -1,0 +1,2 @@
+//`./api/`는 따로 관리(@/hooks/api)
+export * from "./useWindows";
